refactor: improve XML handling and URL utilities
- Change fetchXml return type from string to XMLDocument
- Move getParentPath utility to shared/url-utils module
- Update corresponding unit tests
