+++ src/ogc-api/link-utils.ts
@@ -1,6 +1,7 @@
 import { OgcApiDocument, OgcApiDocumentLink } from './model.js';
 import { EndpointError } from '../shared/errors.js';
 import { sharedFetch } from '../shared/http-utils.js';
+import { getParentPath } from '../shared/url-utils.js';
 
 export function fetchDocument<T extends OgcApiDocument>(
   url: string
@@ -134,14 +135,4 @@ export function assertHasLinks(
 ) {
   if (!hasLinks(doc, relType))
     throw new EndpointError(`Could not find link with type: ${relType}`);
-}
-
-export function getParentPath(url: string): string | null {
-  const urlObj = new URL(url, window.location.toString());
-  const pathParts = urlObj.pathname.replace(/\/$/, '').split('/');
-  if (pathParts.length <= 2) {
-    return null;
-  }
-  urlObj.pathname = pathParts.slice(0, -1).join('/');
-  return urlObj.toString();
 }
